hs9-flask-api-raamatute-otsing/requirements.txt
flask==3.0.2
flask-cors
requests==2.31.0
azure-storage-blob